[REF] Move nn.Module dtype to test case definition, not setup_case
Add .double() to Sequential() definitions in all test case lists
(SIMPLIFY_CASES, LAPLACIAN_CASES, BILAPLACIAN_CASES, EXP01_CASES)
and remove redundant f.double() from setup_case and test_jet.
Fix stale report_nonclose imports (now in test.utils).
Revert laplacian.py weighting refactor, keep only jet call update.

+++ test/exp/test_exp01_benchmark_laplacian.py
@@ -17,7 +17,7 @@ from jet.exp.exp01_benchmark_laplacian.execute import (
 from jet.laplacian import SUPPORTED_DISTRIBUTIONS as LAPLACIAN_SUPPORTED_DISTRIBUTIONS
 from jet.utils import run_seeded
 from jet.weighted_laplacian import get_weighting
-from test.test___init__ import report_nonclose, setup_case
+from test.test___init__ import setup_case
 from test.test_bilaplacian import bilaplacian
 from test.test_laplacian import (
     WEIGHT_IDS,
@@ -26,6 +26,7 @@ from test.test_laplacian import (
     get_coefficients,
     laplacian,
 )
+from test.utils import report_nonclose
 
 STRATEGY_IDS = [f"strategy={s}" for s in SUPPORTED_STRATEGIES]
 LAPLACIAN_DISTRIBUTION_IDS = [
@@ -43,7 +44,7 @@ EXP01_CASES = [
     {
         "f": Sequential(
             Linear(5, 4, bias=False), Tanh(), Linear(4, 1, bias=True), Tanh()
-        ),
+        ).double(),
         "mock_args_fn": lambda: (rand(5).double(),),
         "id": "two-layer-tanh-mlp",
     },
